+++ neo/d3xp/gamesys/Class.h
@@ -258,8 +258,8 @@ private:
 	static idList<idTypeInfo *, TAG_IDCLASS>	types;
 	static idList<idTypeInfo *, TAG_IDCLASS>	typenums;
 	static int					typeNumBits;
-	static int					memused;
-	static int					numobjects;
+	static size_t				memused;
+	static size_t				numobjects;
 };
 
 /***********************************************************************
