+++ neo/d3xp/gamesys/Event.h
@@ -58,7 +58,7 @@ private:
 	int							returnType;
 	int							numargs;
 	size_t						argsize;
-	int							argOffset[ D_EVENT_MAXARGS ];
+	size_t						argOffset[ D_EVENT_MAXARGS ];
 	int							eventnum;
 	const idEventDef *			next;
 
@@ -74,8 +74,8 @@ public:
 	char						GetReturnType() const;
 	int							GetEventNum() const;
 	int							GetNumArgs() const;
-	size_t						GetArgSize() const;
-	int							GetArgOffset( int arg ) const;
+	int							GetArgSize() const;
+	size_t						GetArgOffset( int arg ) const;
 
 	static int					NumEventCommands();
 	static const idEventDef		*GetEventCommand( int eventnum );
@@ -184,8 +184,8 @@ ID_INLINE int idEventDef::GetNumArgs() const {
 idEventDef::GetArgSize
 ================
 */
-ID_INLINE size_t idEventDef::GetArgSize() const {
-	return argsize;
+ID_INLINE int idEventDef::GetArgSize() const {
+	return static_cast<int>( argsize );
 }
 
 /*
@@ -193,7 +193,7 @@ ID_INLINE size_t idEventDef::GetArgSize() const {
 idEventDef::GetArgOffset
 ================
 */
-ID_INLINE int idEventDef::GetArgOffset( int arg ) const {
+ID_INLINE size_t idEventDef::GetArgOffset( int arg ) const {
 	assert( ( arg >= 0 ) && ( arg < D_EVENT_MAXARGS ) );
 	return argOffset[ arg ];
 }
